Reorg simulator files
