+++ tests/unit_test/global_scheduler/test_manager.py
@@ -28,8 +28,8 @@ from llumnix.manager import Manager
 from llumnix.instance_info import InstanceInfo, InstanceLoadCalculator
 from llumnix.server_info import ServerInfo
 from llumnix.queue.queue_type import QueueType
-from llumnix.instance_info import InstanceType
-from llumnix.backends.vllm.simulator import BackendSimVLLM
+from llumnix.global_scheduler.scaling_scheduler import InstanceType
+from llumnix.backends.vllm.sim_llm_engine import BackendSimVLLM
 from llumnix.backends.backend_interface import BackendType
 from llumnix.backends.profiling import LatencyMemData
 from llumnix.entrypoints.utils import LaunchMode
@@ -226,8 +226,8 @@ def test_init_instances(ray_env, manager):
 def test_init_instances_sim(ray_env, manager):
     manager.profiling_result_file_path="//"
     # pylint: disable=import-outside-toplevel
-    import llumnix.backends.vllm.simulator
-    llumnix.backends.vllm.simulator.BackendSimVLLM = MockBackendSim
+    import llumnix.backends.vllm.sim_llm_engine
+    llumnix.backends.vllm.sim_llm_engine.BackendSimVLLM = MockBackendSim
     engine_args = EngineArgs(model="facebook/opt-125m", worker_use_ray=True)
     _, instances = ray.get(manager.init_instances.remote(QueueType("rayqueue"), BackendType.SIM_VLLM, InstanceArgs(), engine_args))
     num_instances = len(instances)
